Refactor App component to update import path for RecipeDetailsPage
