+++ recipe-sharing-app/src/App.jsx
@@ -3,7 +3,7 @@ import Header from './components/Header'
 import HomePage from './pages/HomePage'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import AddRecipePage from './pages/AddRecipePage'
-import RecipeDetailsPage from './components/RecipeDetailsPage'
+import RecipeDetailsPage from './pages/RecipeDetailsPage'
 
 function App() {
